[634215] move test to new location per rlr's suggestion
